+++ modules/hanna-react/src/Selectbox.tsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useRef, useState } from 'react';
 import type {
-  SelectboxOption as _SelectboxOption,
-  SelectboxOptions as _SelectboxOptions,
+  SelectboxOption,
   SelectboxProps as _SelectboxProps,
 } from '@hugsmidjan/react/Selectbox';
 import _Selectbox from '@hugsmidjan/react/Selectbox';
@@ -9,10 +8,12 @@ import getBemClass from '@hugsmidjan/react/utils/getBemClass';
 
 import FormField, { FormFieldWrappingProps } from './FormField';
 
-export type SelectboxOption = _SelectboxOption;
-/** @deprecated Use `SelectboxOptionList` instead  (Will be removed in v0.11) */
-export type SelectboxOptions = _SelectboxOptions;
-export type SelectboxOptionList = _SelectboxOptions;
+export {
+  type SelectboxOption,
+  type SelectboxOptions as SelectboxOptionList,
+  /** @deprecated Use `SelectboxOptionList` instead  (Will be removed in v0.11) */
+  type SelectboxOptions,
+} from '@hugsmidjan/react/Selectbox';
 
 const getValue = (opt: SelectboxOption | string | number | undefined) => {
   const val = typeof opt === 'object' ? opt.value : opt;